pyproject.toml
[tool.poetry]
name = "spotify-client"
version = "0.1.0"
description = ""
authors = ["krisgholson <kris@idea2.dev>"]
readme = "README.md"
packages = [{include = "spotify_client"}]

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
